+++ docs/en/changes/changes.md
@@ -106,6 +106,7 @@
 * BanyanDB: fix when setting `@BanyanDB.TimestampColumn`, the column should not be indexed.
 * OAP Self Observability: make Trace analysis metrics separate by label `protocol`, add Zipkin span dropped metrics.
 * Enhance the alarm kernel with recovered status notification capability
+* BanyanDB: Move data write logic from BanyanDB Java Client to OAP and support observe metrics for write operations.
 
 #### UI